+++ backend/schemas.py
@@ -4,18 +4,18 @@ from pydantic import BaseModel
 class ProductCreate(BaseModel):
     name: str
     quantity: int
-    price: int
+    price: float
 
 class ProductUpdate(BaseModel):
     name: str
     quantity: int
-    price: int
+    price: float
 
 class Product(BaseModel):
     id: int
     name: str
     quantity: int
-    price: int
+    price: float
 
     class Config:
         from_attributes = True
